+++ src/Compilers/CSharp/Portable/Lowering/LambdaRewriter/LambdaRewriter.LocalFunctionReferenceRewriter.cs
@@ -1,11 +1,8 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using Microsoft.CodeAnalysis.CodeGen;
 using Microsoft.CodeAnalysis.CSharp.Symbols;
-using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Diagnostics;
-using System;
 
 namespace Microsoft.CodeAnalysis.CSharp
 {
@@ -81,7 +78,7 @@ namespace Microsoft.CodeAnalysis.CSharp
 
 
         private void RemapLocalFunction(
-            CSharpSyntaxNode syntax,
+            SyntaxNode syntax,
             MethodSymbol symbol,
             out BoundExpression receiver,
             out MethodSymbol method,
