+++ src/Compilers/Core/CodeAnalysisTest/Collections/BitArrayTests.cs
@@ -104,33 +104,50 @@ namespace Microsoft.CodeAnalysis.UnitTests.Collections
         {
             var r = new Random(seed);
             for (int capacity = 0; capacity < maxBits; capacity++)
-                CheckUnionCore(capacity, r);
+            {
+                CheckUnionCore(capacity, capacity, r);
+            }
+
             for (int i = 0; i < rounds; i++)
             {
-                CheckUnionCore(r.Next(maxBits), r);
+                CheckUnionCore(r.Next(maxBits), r.Next(maxBits), r);
             }
         }
 
-        private void CheckUnionCore(int capacity, Random r)
+        private void CheckUnionCore(int capacity1, int capacity2, Random r)
         {
             BitVector b1 = BitVector.Empty, b2 = BitVector.Empty;
-            b1.EnsureCapacity(capacity);
-            b2.EnsureCapacity(capacity);
-            bool[] a1 = new bool[capacity], a2 = new bool[capacity];
-            for (int i = 0; i < capacity; i++)
+            b1.EnsureCapacity(capacity1);
+            b2.EnsureCapacity(capacity2);
+
+            var maxCapacity = Math.Max(capacity1, capacity2);
+            bool[] a1 = new bool[maxCapacity],
+                   a2 = new bool[maxCapacity];
+
+            for (int i = 0; i < capacity1; i++)
             {
                 b1[i] = a1[i] = r.NextBool();
+            }
+
+            for (int i = 0; i < capacity2; i++)
+            {
                 b2[i] = a2[i] = r.NextBool();
             }
-            b1.UnionWith(b2);
-            for (int i = 0; i < capacity; i++)
+
+            bool changed = b1.UnionWith(b2);
+            bool changed2 = false;
+
+            for (int i = 0; i < maxCapacity; i++)
             {
+                bool a = a1[i];
                 a1[i] |= a2[i];
+                changed2 |= (a != a1[i]);
             }
-            for (int i = 0; i < capacity; i++)
+            for (int i = 0; i < maxCapacity; i++)
             {
                 Assert.Equal(a1[i], b1[i]);
             }
+            Assert.Equal(changed2, changed);
         }
 
         [Fact]
